requirements.txt
certifi==2020.6.20
chardet==3.0.4
clarifai==2.6.2
click==7.1.2
configparser==3.8.1
Flask==1.1.2
flask-googlemaps==0.4.0
future==0.18.2
geojson==2.5.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==2.6.0
MarkupSafe==1.1.1
numpy==1.19.1
protobuf==3.12.2
PyJWT==1.7.1
pyngrok==4.1.6
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
reverse-geocoder==1.5.1
scipy==1.5.1
six==1.15.0
twilio==6.44.1
urllib3==1.25.10
Werkzeug==1.0.1
